requirements.txt
yfinance
pandas
seaborn
matplotlib
python-dotenv
